522: v1.0: did check_list and logs
