--- .env_example
@@ -1,20 +1,0 @@
-RECAPTCHA_SITE_KEY=123
-RECAPTCHA_SECRET_KEY=123
-SECRET_KEY=123
-ADMIN_PASSWORD=admin
-SIGNUP_PAGE_ENABLED=False
-
-CONSUMER_KEY=key
-CONSUMER_SECRET=secret
-
-WEB_PORT=8080
-
-REDIS_URL=redis://redis:6379/0
-CELERY_QUERIES=check-solution
-
-FLOWER_SECRET_KEY=123
-FLOWER_PORT=5555
-FLOWER_AUTH=user:user
-FLOWER_PREFIX=/
-
-MONGODB_CACHE_SIZE=1
