+++ app/routes/lti.py
@@ -1,0 +1,15 @@
+# from flask import Blueprint, render_template, jsonify, request, redirect, url_for, abort
+# from app.utils import format_check_for_table, checklist_filter
+# from app.db import db_methods
+# from flask_login import login_required, current_user, logout_user, login_user
+#
+# from app.lti_session_passback.lti import utils
+# from app.lti_session_passback.lti.check_request import check_request
+#
+# from app.main.check_packs import BASE_PACKS, BaseCriterionPack, DEFAULT_TYPE
+# from logs import logger
+# # lti = Blueprint('lti', __name__, template_folder='templates', static_folder='static')
+#
+# from app.server_consts import ALLOWED_EXTENSIONS
+
+
